Common functions

AbstractSDRs.getSupportedSDR — Method

Returns an array of symbol which lists the supported SDR backends

–- Syntax

l = getSupportedSDR()

–- Input parameters

–- Output parameters

l : Array of symbols of supported SDRs

AbstractSDRs.openSDR — Method

Open a Software Defined Radio of backend 'type', tune accordingly based on input parameters and use the supported keywords. It returns a radio object, depending on the type of SDR that can be used with all AbstractSDRs supported functions

–- Syntax

radio = openSDR(type,carrierFreq,samplingRate,gain,antenna;key)

–- Input parameters

type : Desired SDR type. The different supported radio format can be obtained with getSupportedSDR();

carrierFreq : Carrier frequency [Hz]

samplingRate : Sampling frequency (Hz)

gain : Analog Rx gain (dB)

–- Output parameters

radio : Defined SDR object

AbstractSDRs.recv! — Method

Receive from the SDR and fill them in the input buffer.

–- Syntax

nbSamples = recv!(sig,radio);

–- Input parameters

sig : Buffer to be filled

radio : SDR device

–- Output parameters

nbSamples : Number of samples filled

AbstractSDRs.recv — Method

Receive nbSamples from the SDR and fill them in the output buffer. The buffer format depends on the SDR backend

–- Syntax

buffer = recv(radio, nbSamples)

–- Input parameters

radio : SDR object

nbSamples : Desired number of samples

–- Output parameters

buffer : Output buffer from the radio filled with nbSamples samples

AbstractSDRs.updateGain! — Method

Update gain of current radio device, and update radio object with the new obtained gain. If the input is a [UHDRx] or a [UHDTx] object, it updates only the Rx or Tx gain

–- Syntax

updateGain!(radio,gain)

–- Input parameters

radio : SDR device

gain : New desired gain

–- Output parameters

AbstractSDRs.updateSamplingRate! — Method

Update sampling rate of current radio device, and update radio object with the new obtained sampling frequency.

–- Syntax

updateSamplingRate!(radio,samplingRate)

–- Input parameters

radio : SDR device

samplingRate : New desired sampling rate

–- Output parameters